# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CheerConnect API",
    "description": "REST API for CheerConnect — a social network for the cheerleading community. Built with Next.js 16 App Router, PostgreSQL via Prisma ORM, and NextAuth.js for authentication.",
    "version": "1.0.0",
    "contact": {
      "name": "CheerConnect"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth.js JWT session cookie. Obtain by logging in via POST /api/auth/callback/credentials or OAuth flow."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          }
        },
        "required": [
          "error"
        ]
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "Position": {
        "type": "string",
        "enum": [
          "FLYER",
          "BASE",
          "BACKSPOT",
          "FRONTSPOT",
          "TUMBLER",
          "COACH",
          "CHOREOGRAPHER",
          "JUDGE",
          "OTHER"
        ]
      },
      "CareerRole": {
        "type": "string",
        "enum": [
          "ATHLETE",
          "COACH",
          "ASSISTANT_COACH",
          "CHOREOGRAPHER",
          "TEAM_MANAGER",
          "JUDGE",
          "OTHER"
        ]
      },
...
      "TeamSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "logo": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Achievement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string"
          }
        }
      },
...
      "CareerHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/CareerRole"
          },
          "positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Position"
            }
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isCurrent": {
            "type": "boolean"
          },
          "teamName": {
            "type": "string"
          },
          "teamId": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "team": {
            "$ref": "#/components/schemas/TeamSummary"
          }
        }
      },
...
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Não autorizado"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request data",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
...
  "paths": {
...
    "/api/connections/{id}/accept": {
      "post": {
        "tags": [
          "Connections"
        ],
        "summary": "Accept connection request",
        "description": "The path parameter is the sender's user ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Sender's user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Connection accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "connection": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "example": "ACCEPTED"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/career": {
      "get": {
        "tags": [
          "Career"
        ],
        "summary": "List career history",
        "description": "Get the current user's career history entries.",
        "responses": {
          "200": {
            "description": "Career history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "careerHistory": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CareerHistory"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Career"
        ],
        "summary": "Add career entry",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "role",
                  "startDate",
                  "teamName"
                ],
                "properties": {
                  "role": {
                    "$ref": "#/components/schemas/CareerRole"
                  },
                  "positions": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Position"
                    },
                    "maxItems": 10
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date",
                    "nullable": true
                  },
                  "isCurrent": {
                    "type": "boolean",
                    "default": false
                  },
                  "teamName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "teamId": {
                    "type": "string",
                    "nullable": true
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "location": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Career entry created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "career": {
                      "$ref": "#/components/schemas/CareerHistory"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/api/achievements/{id}": {
      "patch": {
        "tags": [
          "Achievements"
        ],
        "summary": "Update achievement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "category": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Achievement updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "achievement": {
                      "$ref": "#/components/schemas/Achievement"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Achievements"
        ],
        "summary": "Delete achievement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Achievement deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}